chore: rename SystemObject and NonStateProp
