+++ Assets/SWPPT3/Scripts/Main/ConditionObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ae13c155d5340c89a85413aba302c33
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
