+++ Assets/SWPPT3/Scripts/Main/ConditionObject/ConditionObjectBase.cs
@@ -1,8 +1,8 @@
 using SWPPT3.Main.Prop;
 
-namespace SWPPT3.Main.SystemObject
+namespace SWPPT3.Main.ConditionObject
 {
-    public abstract class SystemObjectBase
+    public abstract class ConditionObjectBase
     {
         protected abstract void IsSatisfied();
