+++ Assets/SWPPT3/Scripts/Main/ConditionObject/ConditionObjectBase.cs.meta
