+++ Assets/SWPPT3/Scripts/Main/ConditionObject/DoorSystem.cs
@@ -1,0 +1,12 @@
+namespace SWPPT3.Main.ConditionObject
+{
+    public class DoorSystem : ConditionObjectBase
+    {
+        protected override void IsSatisfied()
+        {
+
+        }
+
+
+    }
+}
