+++ Assets/SWPPT3/Scripts/Main/ConditionObject/DoorSystem.cs.meta
