+++ Assets/SWPPT3/Scripts/Main/Main.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b34e3d0affda1c04299dc3e54fea19b9
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
