+++ Assets/SWPPT3/Scripts/Main/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc6a4948f38a5af49a18ef339467a999
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
