+++ Assets/SWPPT3/Scripts/Main/Player/MetalState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3169da9b00704d5ca8af92145b818f11
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
