+++ Assets/SWPPT3/Scripts/Main/Player/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c834ef87520fa146a131b8e2e184ab9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
