+++ Assets/SWPPT3/Scripts/Main/Player/PlayerState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ab3ad21b92d4122b7a25f1b91addf95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
