+++ Assets/SWPPT3/Scripts/Main/Player/RubberState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 657e97c2a5b64e11b6a2e907fdf7e469
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
