+++ Assets/SWPPT3/Scripts/Main/Player/SlimeState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddbaea36c56443d784f608868a962a52
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
