+++ Assets/SWPPT3/Scripts/Main/PlayerMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b2daab0e2304068b9b6802f0836616b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
