+++ Assets/SWPPT3/Scripts/Main/Prop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee5c5748a274f794bb8e7ec01f96b76a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
