+++ Assets/SWPPT3/Scripts/Main/Prop/ElectricPool.cs
@@ -1,6 +1,6 @@
 namespace SWPPT3.Main.Prop
 {
-    public class ElectricPool : NonStatePropBase
+    public class ElectricPool : StatelessPropBase
     {
 
     }
