+++ Assets/SWPPT3/Scripts/Main/Prop/ElectricPool.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d9974bae7c24fa3bf55c032fc43f2a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
