+++ Assets/SWPPT3/Scripts/Main/Prop/ElectricWire.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 720894524c224583b5123c0737417b54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
