+++ Assets/SWPPT3/Scripts/Main/Prop/FloorButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7ceeb03abe9e4947a34f2e13180de72e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
