+++ Assets/SWPPT3/Scripts/Main/Prop/ItemBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bba6f14c6f242cea6b0b0e510071188
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
