--- Assets/SWPPT3/Scripts/Main/Prop/NonStatePropBase.cs
@@ -1,7 +1,0 @@
-namespace SWPPT3.Main.Prop
-{
-    public abstract class NonStatePropBase : PropBase
-    {
-
-    }
-}
