+++ Assets/SWPPT3/Scripts/Main/Prop/PoisonPool.cs
@@ -1,6 +1,6 @@
 namespace SWPPT3.Main.Prop
 {
-    public class PoisonPool : NonStatePropBase
+    public class PoisonPool : StatelessPropBase
     {
 
     }
