+++ Assets/SWPPT3/Scripts/Main/Prop/PoisonPool.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8981e25f7d9e469b851d17f403c134ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
