+++ Assets/SWPPT3/Scripts/Main/Prop/PropBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d00a654da9e7042458940bd73ea88fc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
