+++ Assets/SWPPT3/Scripts/Main/Prop/StateProps.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2511cfb6c7c4454cae806380882a4e2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
