+++ Assets/SWPPT3/Scripts/Main/Prop/StatelessPropBase.cs
@@ -1,0 +1,7 @@
+namespace SWPPT3.Main.Prop
+{
+    public abstract class StatelessPropBase : PropBase
+    {
+
+    }
+}
