+++ Assets/SWPPT3/Scripts/Main/Prop/StatelessPropBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3110faa25396bf41ba93a541754f96c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
