+++ Assets/SWPPT3/Scripts/Main/Prop/WoodBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 049246bd5088433cbf69b64e1b7309ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
