--- Assets/SWPPT3/Scripts/Main/SystemObject/DoorSystem.cs
@@ -1,12 +1,0 @@
-namespace SWPPT3.Main.SystemObject
-{
-    public class DoorSystem : SystemObjectBase
-    {
-        protected override void IsSatisfied()
-        {
-
-        }
-
-
-    }
-}
