+++ Assets/SWPPT3/Scripts/Main/YouSpinMeRound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d247ef9c3f335904dad3cda0a1b25136
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
